Refactoring - model classes into model file, utility function into util
