+++ models/utils/util.py
@@ -1,0 +1,39 @@
+import tensorflow as tf
+import numpy as np
+from PIL import Image
+import matplotlib.pyplot as plt
+
+
+def load_image(image_path):
+    img = tf.io.read_file(image_path)
+    img = tf.image.decode_jpeg(img, channels=3)
+    img = tf.image.resize(img, (299, 299))
+    img = tf.keras.applications.inception_v3.preprocess_input(img)
+    return img, image_path
+
+def get_image_feature_extractor():
+
+    image_model = tf.keras.applications.InceptionV3(include_top=False,
+                                                    weights='imagenet')
+    new_input = image_model.input
+    hidden_layer = image_model.layers[-1].output
+
+    return tf.keras.Model(new_input, hidden_layer)
+
+
+def plot_attention(image, result, attention_plot):
+    temp_image = np.array(Image.open(image))
+
+    fig = plt.figure(figsize=(10, 10))
+
+    len_result = len(result)
+    for i in range(len_result):
+        temp_att = np.resize(attention_plot[i], (8, 8))
+        grid_size = max(np.ceil(len_result / 2), 2)
+        ax = fig.add_subplot(grid_size, grid_size, i + 1)
+        ax.set_title(result[i])
+        img = ax.imshow(temp_image)
+        ax.imshow(temp_att, cmap='gray', alpha=0.6, extent=img.get_extent())
+
+    plt.tight_layout()
+    plt.show()
