+++ models/vqa/naive_vqa.py
@@ -1,0 +1,2 @@
+import tensorflow as tf
+
